hoàn thành version 1

+++ src/main/java/com/app/lms/service/AnswerOptionService.java
@@ -58,6 +58,9 @@ public class AnswerOptionService {
     }
 
     public void deleteAnswerOption (Long answerOptionId) {
+        if (!answerOptionRepository.existsById(answerOptionId)) {
+            throw new AppException(ErroCode.ANSWER_OPTION_NO_EXISTED);
+        }
         answerOptionRepository.deleteById(answerOptionId);
     }
